extract servers.go for all kinds of servers
